texture for cable_drum and move dummy block classes

+++ src/main/java/com/hbm/blocks/machine/dummy/DummyBlockRefinery.java
@@ -1,13 +1,13 @@
-package com.hbm.blocks.machine;
+package com.hbm.blocks.machine.dummy;
 
 import java.util.Random;
 
 import com.hbm.blocks.ModBlocks;
 import com.hbm.main.MainRegistry;
 import com.hbm.tileentity.machine.TileEntityDummy;
-import com.hbm.tileentity.machine.TileEntityDummyPortNew;
+import com.hbm.tileentity.machine.TileEntityDummyFluidPort;
+import com.hbm.tileentity.machine.oil.TileEntityMachineRefinery;
 
-import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.player.EntityPlayer;
@@ -18,77 +18,53 @@ import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumBlockRenderType;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
-import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 
-public class DummyBlockMachine extends DummyOldBase {
+public class DummyBlockRefinery extends DummyOldBase {
 
 	public static boolean safeBreak = false;
-	
-	private int id;
-	private Block drop;
-	float oX = 0;
-	float oY = 0;
-	float oZ = 0;
-	float dX = 1;
-	float dY = 1;
-	float dZ = 1;
-	public AxisAlignedBB dummy_BB = FULL_BLOCK_AABB;
-	
-	public DummyBlockMachine(Material materialIn, String s, boolean port) {
-		super(materialIn, s, port);
-	}
-	
-	public DummyBlockMachine(Material materialIn, String s, boolean port, int id, Block drop) {
-		super(materialIn, s, port);
-		this.id = id;
-		this.drop = drop;
-	}
-	
-	public DummyBlockMachine setBounds(float oX, float oY, float oZ, float dX, float dY, float dZ) {
 
-		oX *= 0.0625F;
-		oY *= 0.0625F;
-		oZ *= 0.0625F;
-		dX *= 0.0625F;
-		dY *= 0.0625F;
-		dZ *= 0.0625F;
-		this.dummy_BB = new AxisAlignedBB(oX, oY, oZ, dX, dY, dZ);
-		return this;
+	public DummyBlockRefinery(Material materialIn, String s, boolean port) {
+		super(materialIn, s, port);
 	}
 
 	@Override
 	public TileEntity createNewTileEntity(World worldIn, int meta) {
-		return port ? new TileEntityDummyPortNew() : new TileEntityDummy();
+		if(this == ModBlocks.dummy_port_refinery){
+			return new TileEntityDummyFluidPort();
+		} else {
+			return new TileEntityDummy();
+		}
 	}
 	
 	@Override
 	public void breakBlock(World world, BlockPos pos, IBlockState state) {
+		if(!safeBreak) {
     		TileEntity te = world.getTileEntity(pos);
     		if(te != null && te instanceof TileEntityDummy) {
-    		
     			if(!world.isRemote)
     				world.destroyBlock(((TileEntityDummy)te).target, true);
     		}
+    	}
     	world.removeTileEntity(pos);
+	}
+	
+	@Override
+	public EnumBlockRenderType getRenderType(IBlockState state) {
+		return EnumBlockRenderType.INVISIBLE;
+	}
+	
+	@Override
+	public Item getItemDropped(IBlockState state, Random rand, int fortune) {
+		return Items.AIR;
 	}
 	
 	@Override
 	public boolean isOpaqueCube(IBlockState state) {
 		return false;
-	}
-	
-	@Override
-	public boolean isFullCube(IBlockState state) {
-		return false;
-	}
-
-	@Override
-	public EnumBlockRenderType getRenderType(IBlockState state) {
-		return EnumBlockRenderType.INVISIBLE;
 	}
 	
 	@Override
@@ -109,14 +85,10 @@ public class DummyBlockMachine extends DummyOldBase {
 	public boolean shouldSideBeRendered(IBlockState blockState, IBlockAccess blockAccess, BlockPos pos, EnumFacing side) {
 		return false;
 	}
-	@Override
-	public Item getItemDropped(IBlockState state, Random rand, int fortune) {
-		return Items.AIR;
-	}
 	
 	@Override
 	public ItemStack getPickBlock(IBlockState state, RayTraceResult target, World world, BlockPos pos, EntityPlayer player) {
-		return new ItemStack(Item.getItemFromBlock(drop));
+		return new ItemStack(ModBlocks.machine_refinery);
 	}
 	
 	@Override
@@ -128,25 +100,19 @@ public class DummyBlockMachine extends DummyOldBase {
 		{
     		TileEntity te = world.getTileEntity(pos);
     		if(te != null && te instanceof TileEntityDummy) {
-    			int a = ((TileEntityDummy)te).target.getX();
-    			int b = ((TileEntityDummy)te).target.getY();
-    			int c = ((TileEntityDummy)te).target.getZ();
+
+    			BlockPos target = ((TileEntityDummy)te).target;
     			
-    			if(te != null)
+    			TileEntityMachineRefinery entity = (TileEntityMachineRefinery) world.getTileEntity(target);
+    			if(entity != null)
     			{
-    				player.openGui(MainRegistry.instance, id, world, a, b, c);
+    				player.openGui(MainRegistry.instance, ModBlocks.guiID_machine_refinery, world, target.getX(), target.getY(), target.getZ());
     			}
     		}
 			return true;
 		} else {
 			return false;
 		}
-	}
-	
-	
-	@Override
-	public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos) {
-		return this.dummy_BB;
 	}
 
 }
